Add freeze points with friends
