+++ freezing/web/templates/alt_scoring/indiv_hains.html
@@ -1,22 +1,22 @@
 {% extends "base.html" %}
 {% block content %}
+    <h2 class="mb-3">
+        Hains Point Laps
+    </h2>
+    <p class="lead">
+        We're looking for the hamsterest hamster who ever hamstered.
+    </p>
+    <p>
+        Segment Miles is based on the 2.5 mile <a href="https://www.strava.com/segments/1081507">Gate
+        to Gate segment</a>, so it undercounts your mileage if you did a full lap.
+        However, since this segment is the official measure of "A Hains Point Lap" as far as
+        this contest is concerned, it was thought best to use 2.5 miles.
+    </p>
+    <p>
+        If you feel the wrong decision has been made here, please be reminded that you are a hamster, and hamsters don't get the admin password to the web server.
+    </p>
     <div class="card">
-        <h5 class="card-header">
-            Hains Point Laps
-        </h5>
         <div class="card-body">
-            <p class="lead">
-                We're looking for the hamsterest hamster who ever hamstered.
-            </p>
-            <p>
-                Segment Miles is based on the 2.5 mile <a href="https://www.strava.com/segments/1081507">Gate
-                to Gate segment</a>, so it undercounts your mileage if you did a full lap.
-                However, since this segment is the official measure of "A Hains Point Lap" as far as
-                this contest is concerned, it was thought best to use 2.5 miles.
-            </p>
-            <p>
-                If you feel the wrong decision has been made here, please be reminded that you are a hamster, and hamsters don't get the admin password to the web server.
-            </p>
             <div class="table-responsive">
                 <table id="hamster" class="table table-condensed">
                     <thead>
