+++ freezing/web/templates/alt_scoring/team_hains.html
@@ -1,34 +1,34 @@
 {% extends "base.html" %}
 {% block content %}
+    <h2 class="mb-3">
+        Team Hains Point Laps
+    </h2>
+    <p>
+        Certainly hope at least some of these were together as a team.
+    </p>
     <div class="card">
-        <h5 class="card-header">
-            Team Hains Point Laps
-        </h5>
-        <div class="card-body">
-            <p>
-                Certainly hope at least some of these were together as a team.
-            </p>
+        <div class="card-body py-1">
             <table class="table table-condensed">
                 <tr>
-                    <th>
+                    <th class="number rank">
                         Rank
                     </th>
                     <th>
                         Team
                     </th>
-                    <th>
+                    <th class="number">
                         Laps
                     </th>
                 </tr>
                 {% for a, b in team_hains %}
                     <tr>
-                        <td>
+                        <td class="number rank">
                             {{ loop.index }}
                         </td>
                         <td>
                             {{ a }}
                         </td>
-                        <td>
+                        <td class="number">
                             {{ b }}
                         </td>
                     </tr>
