+++ freezing/web/templates/alt_scoring/team_sleaze.html
@@ -1,37 +1,37 @@
 {% extends "base.html" %}
 {% block content %}
+    <h2 class="mb-3">
+        Team Sleaze
+    </h2>
+    <p>
+        Now, you know it's up to you whether or not you want to just do the bare minimum. People can go for a bike ride anywhere, okay? They do Freezing Saddles for the camaraderie and the soul-crushing cold.
+    </p>
+    <p>
+        Look, we want you to ride your bike, okay? Now if you feel that the bare minimum is enough, then okay. But some people choose to ride more and we encourage that, okay? You do want to ride your bike, don't you?
+    </p>
     <div class="card">
-        <h5 class="card-header">
-            Team Sleaze
-        </h5>
-        <div class="card-body">
-            <p>
-                Now, you know it's up to you whether or not you want to just do the bare minimum. People can go for a bike ride anywhere, okay? They do Freezing Saddles for the camaraderie and the soul-crushing cold.
-            </p>
-            <p>
-                Look, we want you to ride your bike, okay? Now if you feel that the bare minimum is enough, then okay. But some people choose to ride more and we encourage that, okay? You do want to ride your bike, don't you?
-            </p>
+        <div class="card-body py-1">
             <table class="table table-condensed">
                 <tr>
-                    <th>
+                    <th class="number rank">
                         Rank
                     </th>
                     <th>
                         Team
                     </th>
-                    <th>
+                    <th class="number">
                         Sleaze Rides
                     </th>
                 </tr>
                 {% for a, b in team_sleaze %}
                     <tr>
-                        <td>
+                        <td class="number rank">
                             {{ loop.index }}
                         </td>
                         <td>
                             {{ a }}
                         </td>
-                        <td>
+                        <td class="number">
                             {{ b }}
                         </td>
                     </tr>
