+++ freezing/web/templates/base.html
@@ -118,13 +118,13 @@
                                     <hr class="dropdown-divider">
                                 </li>
                                 <li>
-                                    <a class="dropdown-item" href="/alt_scoring/team_hains">Team Hains Point Laps</a>
+                                    <a class="dropdown-item" href="/leaderboard/team_hains">Team Hains Point Laps</a>
                                 </li>
                                 <li>
                                     <a class="dropdown-item" href="/leaderboard/team_various">Team Pointless (Various)</a>
                                 </li>
                                 <li>
-                                    <a class="dropdown-item" href="/alt_scoring/team_sleaze">Team Sleaze Rides</a>
+                                    <a class="dropdown-item" href="/leaderboard/team_sleaze">Team Sleaze Rides</a>
                                 </li>
                                 <li>
                                     <hr class="dropdown-divider">
@@ -139,16 +139,16 @@
                                     <hr class="dropdown-divider">
                                 </li>
                                 <li>
-                                    <a class="dropdown-item" href="/alt_scoring/indiv_hains">Individual Hains Point Laps</a>
+                                    <a class="dropdown-item" href="/leaderboard/indiv_hains">Individual Hains Point Laps</a>
                                 </li>
                                 <li>
-                                    <a class="dropdown-item" href="/alt_scoring/indiv_freeze">Individual Freeze Points</a>
+                                    <a class="dropdown-item" href="/leaderboard/indiv_freeze">Individual Freeze Points</a>
                                 </li>
                                 <li>
                                     <a class="dropdown-item" href="/leaderboard/individual_various">Individual Pointless (Various)</a>
                                 </li>
                                 <li>
-                                    <a class="dropdown-item" href="/alt_scoring/indiv_sleaze">Individual Sleaze Rides</a>
+                                    <a class="dropdown-item" href="/leaderboard/indiv_sleaze">Individual Sleaze Rides</a>
                                 </li>
                                 <li>
                                     <hr class="dropdown-divider">
