+++ freezing/web/templates/index.html
@@ -100,7 +100,7 @@ body.map-expanded {
                     <div class="modal-body d-flex flex-column align-items-center ekko-lightbox-container">
                         {# djlint:off H006 #}
                         <img src="{{ photo.img_l }}"
-                             class="img-responsive center-block"
+                             class="img-responsive"
                              alt="{{ photo.caption }}">
                         {# djlint:on #}
                     </div>
