+++ freezing/web/templates/leaderboard/indiv_various.html
@@ -83,9 +83,9 @@
     </script>
 {% endblock %}
 {% block content %}
-    <h1>
+    <h2 class="mb-3">
         Individual Leaderboards
-    </h1>
+    </h2>
     <form action="#" id="choose_api" class="d-flex">
         <select class="form-select" name="api_method" id="api_method">
             <option value="indiv_elev_gain" selected="selected">
