+++ freezing/web/templates/leaderboard/team_various.html
@@ -81,9 +81,9 @@
     </script>
 {% endblock %}
 {% block content %}
-    <h1>
+    <h2 class="mb-3">
         Team Leaderboards
-    </h1>
+    </h2>
     <form action="#" id="choose_api" class="d-flex">
         <select class="form-select" name="api_method" id="api_method">
             <option value="team_elev_gain" selected="selected">
